PY-105 Move alpha.exceptions out of alpha
As exceptions will be most probably shared between versions of the API
and are raised from version-independent internal code, I'm moving
exceptions to top-level: `neptune_fetcher.alpha.exceptions` becomes
`neptune_fetcher.exceptions`. Old name is aliased to the new name for
compatibilty with existing scripts.
