+++ src/neptune_fetcher/internal/composition/type_inference.py
@@ -24,8 +24,8 @@ from typing import (
 from neptune_api.client import AuthenticatedClient
 
 from neptune_fetcher.alpha import filters
-from neptune_fetcher.alpha.exceptions import AttributeTypeInferenceError
 from neptune_fetcher.internal import identifiers
+from neptune_fetcher.exceptions import AttributeTypeInferenceError
 from neptune_fetcher.internal.composition import attribute_components as _components
 from neptune_fetcher.internal.composition import concurrency
 from neptune_fetcher.internal.retrieval import attribute_definitions as att_defs
