+++ src/neptune_fetcher/internal/retrieval/util.py
@@ -33,8 +33,8 @@ from neptune_api import AuthenticatedClient
 from neptune_api.errors import ApiKeyRejectedError
 from neptune_retrieval_api.types import Response
 
-from neptune_fetcher.alpha import exceptions
-from neptune_fetcher.alpha.exceptions import (
+from neptune_fetcher import exceptions
+from neptune_fetcher.exceptions import (
     NeptuneInvalidCredentialsError,
     NeptuneProjectInaccessible,
 )
