+++ tests/e2e/alpha/internal/composition/test_type_inference.py
@@ -6,7 +6,7 @@ from datetime import (
 
 import pytest
 
-from neptune_fetcher.alpha.exceptions import AttributeTypeInferenceError
+from neptune_fetcher.exceptions import AttributeTypeInferenceError
 from neptune_fetcher.alpha.filters import (
     Attribute,
     Filter,
